Move EmbeddedTestConfig to config package
