+++ repository/src/test/java/com/epam/esm/repository/impl/CertificateRepositoryImplTest.java
@@ -3,7 +3,7 @@ package com.epam.esm.repository.impl;
 import com.epam.esm.entity.Certificate;
 import com.epam.esm.entity.Tag;
 import com.epam.esm.repository.CertificateRepository;
-import com.epam.esm.repository.EmbeddedTestConfig;
+import com.epam.esm.repository.config.EmbeddedTestConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
