+++ repository/src/test/java/com/epam/esm/repository/impl/SearchRepositoryImplTest.java
@@ -2,8 +2,8 @@ package com.epam.esm.repository.impl;
 
 import com.epam.esm.entity.Certificate;
 import com.epam.esm.entity.ParameterWrapper;
-import com.epam.esm.repository.EmbeddedTestConfig;
 import com.epam.esm.repository.SearchRepository;
+import com.epam.esm.repository.config.EmbeddedTestConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
