+++ repository/src/test/java/com/epam/esm/repository/impl/TagRepositoryImplTest.java
@@ -1,7 +1,7 @@
 package com.epam.esm.repository.impl;
 
 import com.epam.esm.entity.Tag;
-import com.epam.esm.repository.EmbeddedTestConfig;
+import com.epam.esm.repository.config.EmbeddedTestConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
